requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
pytesseract==0.3.10
opencv-python==4.8.1.78
numpy==1.24.3
Pillow==10.1.0
openpyxl==3.1.2
python-dotenv==1.0.0
SQLAlchemy==2.0.23
werkzeug==3.0.1
pandas==2.1.3
gunicorn==21.2.0
pdf2image==1.16.3
PyPDF2==3.0.1
poppler-utils==0.1.0